fixed #944 struct ValueStopWatch instead of Stopwatch in Metrics, Marker -> struct move StatsReport, TypeStat etc to separate file
