+++ src/GraphQL/Instrumentation/ApolloTracingExtensions.cs
@@ -28,20 +28,26 @@ namespace GraphQL.Instrumentation
             PerfRecord[] perf,
             DateTime start)
         {
-            var operationStat = perf.Single(x => x.Category == "operation");
-            var documentStats = perf.Where(x => x.Category == "document");
-            var fieldStats = perf.Where(x => x.Category == "field");
-
+            var operationStat = perf.Single(x => x.Category == "operation"); // always exists
             var trace = new ApolloTrace(start, operationStat.Duration);
 
-            var parsingStat = documentStats.Single(x => x.Subject == "Building document");
-            trace.Parsing.StartOffset = ApolloTrace.ConvertTime(parsingStat.Start);
-            trace.Parsing.Duration = ApolloTrace.ConvertTime(parsingStat.Duration);
+            var documentStats = perf.Where(x => x.Category == "document");
 
-            var validationStat = documentStats.Single(x => x.Subject == "Validating document");
-            trace.Validation.StartOffset = ApolloTrace.ConvertTime(validationStat.Start);
-            trace.Validation.Duration = ApolloTrace.ConvertTime(validationStat.Duration);
+            var parsingStat = documentStats.FirstOrDefault(x => x.Subject == "Building document");
+            if (parsingStat != null) // can be null if exception occured
+            {
+                trace.Parsing.StartOffset = ApolloTrace.ConvertTime(parsingStat.Start);
+                trace.Parsing.Duration = ApolloTrace.ConvertTime(parsingStat.Duration);
+            }
 
+            var validationStat = documentStats.FirstOrDefault(x => x.Subject == "Validating document");
+            if (validationStat != null) // can be null if exception occured
+            {
+                trace.Validation.StartOffset = ApolloTrace.ConvertTime(validationStat.Start);
+                trace.Validation.Duration = ApolloTrace.ConvertTime(validationStat.Duration);
+            }
+
+            var fieldStats = perf.Where(x => x.Category == "field");
             foreach (var fieldStat in fieldStats)
             {
                 var stringPath = fieldStat.MetaField<IEnumerable<string>>("path");
